requirements.txt
Django==4.0.4
python-decouple==3.6
pytest-django==4.5.2
factory-boy==3.2.1
djangorestframework==3.13.1
django-filter==21.1
pytest-cov==3.0.0